requirements.txt
python-dotenv==0.19.*
python-telegram-bot==13.7
redis==3.5.*
vk-api==11.9.*
